String Introduction Hacker Rank
There are three lines of output: 
For the first line, sum the lengths of  and . 
For the second line, write Yes if  is lexicographically greater than  otherwise print No instead. 
For the third line, capitalize the first letter in both  and  and print them on a single line, separated by a space.